Second Commit

+++ src/main/java/com/example/codesimple/entity/Book.java
@@ -8,7 +8,7 @@ public class Book {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String name;
     @Column(name = "description")
     private String desc;
@@ -23,7 +23,7 @@ public class Book {
 
     }
 
-    public Book(Integer id, String name, String desc, Integer yearOfPublication, String bookType) {
+    public Book(Long id, String name, String desc, Integer yearOfPublication, String bookType) {
         this.id = id;
         this.name = name;
         this.desc = desc;
@@ -31,11 +31,11 @@ public class Book {
         this.bookType = bookType;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
